feat: voting buttons and select neurons new design (#1313)
# Motivation

Implement the new UX/UI design approach for the voting actions in proposal details.

# Note

This PR does **not** provide any new tests because there isn't any new feautres. It's only component moving and styles.

# Changes

- refactoring split features in various components (either move or copy/paste) to support "modern" and "legacy" design
- use gix-component bottom-sheet, new icons and new styles
- lost of styling and spacing
