+++ frontend/src/App.svelte
@@ -70,4 +70,5 @@
 <style lang="scss" global>
   @import "@dfinity/gix-components/styles/global.scss";
   @import "lib/themes/legacy";
+  @import "lib/themes/variables";
 </style>
