+++ frontend/src/lib/components/ui/ContentCell.svelte
@@ -7,7 +7,17 @@
 </div>
 
 <style lang="scss">
-  .content-cell-details:not(:last-of-type) {
-    margin-bottom: var(--row-gap);
+  @use "@dfinity/gix-components/styles/mixins/media";
+
+  .content-cell-details:last-of-type {
+    margin-bottom: 0;
+  }
+
+  .content-cell-details {
+    margin-bottom: 0;
+
+    @include media.min-width(large) {
+      margin-bottom: var(--row-gap);
+    }
   }
 </style>
