+++ frontend/src/lib/i18n/en.json
@@ -402,7 +402,7 @@
   "proposal_detail__vote": {
     "headline": "Cast Vote",
     "neurons": "Neurons",
-    "voting_power": "Voting power",
+    "voting_power": "Voting Power",
     "vote_progress": "Vote Progress",
     "total": "Total",
     "adopt": "Adopt",
