+++ frontend/src/lib/themes/variables.scss
@@ -1,0 +1,5 @@
+// NNS-dapp global specific theme variables
+:root {
+  --voting-bottom-sheet-header-height: 102px;
+  --voting-bottom-sheet-content-max-height: 30vh;
+}
