+++ frontend/src/tests/lib/components/proposal-detail/VotingCard/VotingCard.spec.ts
@@ -13,7 +13,6 @@ import { neuronsStore } from "../../../../../lib/stores/neurons.store";
 import { votingNeuronSelectStore } from "../../../../../lib/stores/proposals.store";
 import { mockAuthStoreSubscribe } from "../../../../mocks/auth.store.mock";
 import { MockGovernanceCanister } from "../../../../mocks/governance.canister.mock";
-import en from "../../../../mocks/i18n.mock";
 import { mockNeuron } from "../../../../mocks/neurons.mock";
 import { mockProposalInfo } from "../../../../mocks/proposal.mock";
 import VotingCardTest from "./VotingCardTest.svelte";
@@ -59,9 +58,9 @@ describe("VotingCard", () => {
 
   it("should be visible if there are some not-voted-neurons", async () => {
     neuronsStore.setNeurons({ neurons, certified: true });
-    const { queryByText } = renderVotingCard();
+    const { getByTestId } = renderVotingCard();
     await waitFor(() =>
-      expect(queryByText(en.proposal_detail__vote.headline)).toBeInTheDocument()
+      expect(getByTestId("voting-confirmation-toolbar")).toBeInTheDocument()
     );
   });
