+++ frontend/src/tests/lib/components/proposal-detail/VotingCard/VotingConfirmationToolbar.spec.ts
@@ -25,6 +25,7 @@ describe("VotingConfirmationToolbar", () => {
 
   const props = {
     proposalInfo: mockProposalInfo,
+    layout: "modern",
   };
 
   beforeEach(() => {
@@ -139,7 +140,10 @@ describe("VotingConfirmationToolbar", () => {
 
   it("should display a question that repeats id and topic", async () => {
     const { container } = render(VotingConfirmationToolbar, {
-      props,
+      props: {
+        ...props,
+        layout: "legacy",
+      },
     });
 
     const testLabel = replacePlaceholders(
